requirements.txt
python-telegram-bot>=13.5
pyTelegramBotAPI>=0.0.1
pytz
